requirements.txt
flask
flask_sqlalchemy
flask_migrate
flask_restful
flask_marshmallow